+++ src/lib/stressParser.ts
@@ -25,15 +25,12 @@ export function parseStressWithNoIndex(input: string): Stress[] {
   }))
 }
 
-export function parseStress(
-  input: string,
-  errorMessage = 'validation.stressInvalidNumbersCount'
-): ParsedResult<Stress[]> {
+export function parseStress(input: string): ParsedResult<Stress[]> {
   const isMisesStress = hasValidLineFormat(input, 7)
   const isStressWithoutIndices = hasValidLineFormat(input, 6)
 
   if (isMisesStress) return { data: parseDefaultStress(input) }
   if (isStressWithoutIndices) return { data: parseStressWithNoIndex(input) }
 
-  return { error: { message: errorMessage } }
+  return { error: { message: 'validation.stressInvalidNumbersCount' } }
 }
