+++ src/pages/ModelViewPage.tsx
@@ -4,10 +4,23 @@ import InstrumentsSidebar from '@/components/InstrumentsSidebar'
 import Legend from '@/components/Legend'
 import Scene from '@/components/Scene'
 import Toolbar from '@/components/Toolbar'
-import useLoadHandler from '@/components/useLoadHandler'
 import { useAppDispatch, useAppSelector } from '@/hooks/use-redux'
+import { useModal } from '@/hooks/useModal'
+import { parseCoorinatesMatrix } from '@/lib/coorinatesMatrixParser'
+import { parseOtherCharacteristics } from '@/lib/otherCharacteristicsParser'
+import { parseStress } from '@/lib/stressParser'
+import { buildMisesPhysicalQuantity, calculateMisesStress } from '@/lib/stressUtils'
 import { resetLegend } from '@/redux/slices/legendSlice'
-import { resetModel, setCoorinatesMatrix, setIndicesMatrix, setReady } from '@/redux/slices/modelSlice'
+import {
+  resetModel,
+  setCharacteristic,
+  setCoorinatesMatrix,
+  setDisplacement,
+  setDisplay,
+  setIndicesMatrix,
+  setReady,
+  setStress
+} from '@/redux/slices/modelSlice'
 import { ElementIndices, VertexCoordinate } from '@/types/ModelCommonTypes'
 import { Canvas } from '@react-three/fiber'
 import { useCallback, useMemo, useState } from 'react'
@@ -30,7 +43,8 @@ const ModelViewPage = () => {
     stressMises,
     stressFileName,
     otherCharacteristicFileName,
-    otherCharacteristic
+    otherCharacteristic,
+    stress
   } = useAppSelector((store) => store.model, shallowEqual)
   const dispatch = useAppDispatch()
   const { t } = useTranslation()
@@ -38,6 +52,8 @@ const ModelViewPage = () => {
   const [filesUploaderOpen, setFilesUploaderOpen] = useState(!isReady)
   const [coorinatesMatrixError, setCoorinatesMatrixError] = useState<undefined | string>()
   const [indicesMatrixError, setIndicesMatrixError] = useState<undefined | string>()
+
+  const { openModal } = useModal()
 
   const onModelDelete = useCallback(() => {
     dispatch(resetModel())
@@ -78,7 +94,110 @@ const ModelViewPage = () => {
     setFilesUploaderOpen(false)
   }, [dispatch, indicesMatrix, coorinatesMatrix])
 
-  const { loadCharacteristic, loadStress, loadDisplacement } = useLoadHandler()
+  const loadCharacteristic = useCallback(
+    async (file: File) => {
+      const input = await file.text()
+      const { data: otherCharacteristic, error } = parseOtherCharacteristics(input)
+
+      if (error) {
+        openModal({
+          title: t('validation.error'),
+          message: t(error.message),
+          confirmation: t('validation.checkDataAndTryAgain'),
+          buttons: 'ok'
+        })
+        return
+      }
+
+      if (otherCharacteristic.values.length !== indicesMatrix.length) {
+        openModal({
+          title: t('validation.error'),
+          message: t('validation.otherCharacteristicIsNotTheSameAsElementsCount', {
+            valueCount: otherCharacteristic.values.length,
+            elementsCount: indicesMatrix.length
+          }),
+          confirmation: t('validation.checkDataAndTryAgain'),
+          buttons: 'ok'
+        })
+        return
+      }
+
+      dispatch(setCharacteristic({ otherCharacteristic, fileName: file.name }))
+    },
+    [indicesMatrix, t, dispatch, openModal]
+  )
+
+  const loadStress = useCallback(
+    async (file: File) => {
+      const input = await file.text()
+      const { data: parsedStress, error } = parseStress(input)
+
+      if (error) {
+        openModal({
+          title: t('validation.error'),
+          message: t(error.message),
+          confirmation: t('validation.checkDataAndTryAgain'),
+          buttons: 'ok'
+        })
+        return
+      }
+      if (parsedStress.values.length !== stress.length) {
+        openModal({
+          title: t('validation.error'),
+          message: t('validation.stressIsNotTheSameAsNodesCount', {
+            stressCount: parsedStress.length,
+            elementsCount: stress.length
+          }),
+          confirmation: t('validation.checkDataAndTryAgain'),
+          buttons: 'ok'
+        })
+
+        return
+      }
+
+      const calculatedMises = calculateMisesStress(parsedStress)
+
+      const stressMises = buildMisesPhysicalQuantity(calculatedMises)
+
+      dispatch(setStress({ stressMises, fileName: file.name }))
+      dispatch(setDisplay('stress'))
+    },
+    [dispatch, t, openModal, stress]
+  )
+
+  const loadDisplacement = useCallback(
+    async (file: File) => {
+      const input = await file.text()
+      const { data: displacement, error } = parseCoorinatesMatrix(input, 'validation.displacementInvalidNumbersCount')
+
+      if (error) {
+        openModal({
+          title: t('validation.error'),
+          message: t(error.message),
+          confirmation: t('validation.checkDataAndTryAgain'),
+          buttons: 'ok'
+        })
+        return
+      }
+
+      if (displacement.length !== coorinatesMatrix.length) {
+        openModal({
+          title: t('validation.error'),
+          message: t('validation.displacementIsNotTheSameAsNodesCount', {
+            displacementCount: displacement.length,
+            nodesCount: coorinatesMatrix.length
+          }),
+          confirmation: t('validation.checkDataAndTryAgain'),
+          buttons: 'ok'
+        })
+        return
+      }
+
+      dispatch(setDisplacement({ displacement, displacementFileName: file.name }))
+      dispatch(setDisplay('displacement'))
+    },
+    [coorinatesMatrix, t, dispatch, openModal]
+  )
 
   return (
     <>
